cpu: Add Waiting Instruction Buffer (WIB)
Try to introduce a WIB but
running into probelms with the memory dependence checker. Would need to add structures to grab and realease memory insts without interference in the memory dependence unit.

While trying to figure out this solution a new way to implement the WIB was found. Trying this on a new branch